Use datePickerMode=spinner on Lollipop and higher devices for the DatePicker widget to ensure the puzzle download dialog is displayed correctly on all devices
There was a breaking change in API level 21 (Android 5.0 Lollipop) in which the DatePicker no longer respects the calendarViewShown=false setting.  So now we set both options to ensure that the widget is displayed properly on all devices (note that for tablets or other devices with large screens, the layout-large resource takes precedence).