Fix bug with "stop on end" movement strategy when "skip filled boxes" is enabled
Previously, when filling in the last unfilled letter in word,, the behavior would differ depending on whether or not the current word is the last word in its row or column.  If it was not the last word, the cursor would remain in its prior location, but if it was the last word, then the cursor would move to the last letter in that word.  The behavior is now consistent in all cases and keeps the cursor in its prior location.